# Conflicts:
#	app/ante/fee_calc.go
#	ci-scripts/ansible/roles/decimal-init/templates/genesis-devnet.json.j2
#	ci-scripts/ansible/roles/decimal-init/templates/genesis-fromtestnet.json.j2
#	ci-scripts/ansible/roles/decimal-init/templates/genesis.json.j2
#	proto/decimal/fee/v1/params.proto
#	proto/decimal/nft/v1/query.proto
#	utils/helpers/helpers.go
#	x/coin/errors/errors.go
#	x/coin/keeper/keeper.go
#	x/coin/keeper/msg_server.go
#	x/fee/keeper/market_keeper.go
#	x/fee/types/genesis.go
#	x/fee/types/params.go
#	x/fee/types/params.pb.go

+++ x/multisig/types/constuctors.go
@@ -8,7 +8,7 @@ import (
 )
 
 // NewWallet returns a new Wallet.
-func NewWallet(owners []string, weights []uint64, threshold uint64, salt []byte) (*Wallet, error) {
+func NewWallet(owners []string, weights []uint32, threshold uint32, salt []byte) (*Wallet, error) {
 	// temporary wallet structure to create address
 	w := Wallet{
 		Address:   sdk.AccAddress(salt).String(), // use field Address as salt
